requirements.txt
flask~=3.1.0
flask_cors
flask_sqlalchemy
python-dotenv~=1.0.1
sqlalchemy~=2.0.36

Werkzeug~=3.1.3
pbu~=1.1.10